requirements.txt
streamlit
pandas
matplotlib
sqlalchemy
pytest
plotly
openpyxl
psycopg2-binary
supabase
gotrue
xlsxwriter
fpdf2